# docs/swagger.yml
openapi: 3.0.2
info:
    title: Games_API
    version: 1.0.0
    description: >-
        ### This is a Swagger UI for our Games API  




        ### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a
        more detailed explanation about the error in the body.




        ```

        {
            "error": "Cleartext explanation of the error that occurred"
        }


        ```



        ### Generic API errors


        These error responses can be returned by any of the described endpoints.




        | HTTP status  |  Description |


        |---|---|

        | 400| Bad input parameters or invalid request. More information can be found in the response body.


        | 401| The consumer does not have access to the requested resource or the provided token is not valid.


        | 404| The requested resource can not be found or the user does not have access to the resource.


        | 5XX| A server error has occurred while processing the request.
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    $ref: '#/components/responses/List-of-games'
            summary: Get a list of all games
            description: Returns a list of all games in API
    '/games/{id}':
        summary: Get full details of a game
        description: Returns full details of a given id game
        get:
            tags:
                - details
            responses:
                '200':
                    $ref: '#/components/responses/Full-game-details'
                '404':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                error:
                                    value:
                                        error: error message
                    description: game not found
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        parameters:
            -
                name: id
                schema:
                    type: integer
                in: path
                required: true
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
    responses:
        List-of-games:
            content:
                application/json:
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        List of games:
                            value:
                                - Witcher 3
                                - Cyberpunk 2077
                                - Minecraft
                                - 'Counter-Strike: Global Offensive'
                                - Roblox
                                - Grant Theft Auto V
                                - Valorant
                                - Forza Horizon 5
            description: list of all games
        Full-game-details:
            content:
                application/json:
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        fuul game details:
                            value: |-
                                [
                                    {id:1,name:"Witcher 3",price:29.99},
                                    {id:2,name:"Cyberpunk 2077",price:59.99},
                                    {id:3,name:"Minecraft",price:26.99},
                                    {id:4,name:"Counter-Strike: Global Offensive",price:0},
                                    {id:5,name:"Roblox",price:0},
                                    {id:6,name:"Grant Theft Auto V",price:29.99},
                                    {id:7,name:"Valorant",price:0},
                                    {id:8,name:"Forza Horizon 5",price:59.99}
                                ]
            description: Get full datails of a game